refactor: builtin network
