+++ apps/mobile/src/features/Widget/BuiltInNetwork/SliderTabs/shared.ts
