--- apps/mobile/src/features/Widget/Explorer/internal.ts
@@ -1,4 +1,0 @@
-export type MissionBoxType = {
-	title: string;
-	colors: string[];
-};
