--- apps/mobile/src/features/Widget/components/index.tsx
@@ -1,3 +1,0 @@
-export * from './CollectiblesTab';
-export * from './EmptyTab';
-export * from './TokenTab';
