+++ apps/mobile/src/features/Widget/internal.ts
@@ -1,0 +1,25 @@
+import type { FC } from 'react';
+
+import BuiltInNetwork from './BuiltInNetwork';
+import NotFound from './NotFound';
+import Pixeverse from './Pixeverse';
+import TRexRunner from './TRexRunner';
+
+export interface WidgetProps {
+	id: string;
+}
+
+export type WidgetComponent = FC<WidgetProps>;
+
+export const widgetMap: Record<string, WidgetComponent> = {
+	sui: BuiltInNetwork,
+	tezos: BuiltInNetwork,
+	solana: BuiltInNetwork,
+	aptos: BuiltInNetwork,
+	tRexRunner: TRexRunner,
+	pixeverse: Pixeverse,
+};
+
+export const extractWidgetComponent = (id: string): WidgetComponent => {
+	return widgetMap[id] || NotFound;
+};
