+++ apps/mobile/src/screens/Dashboard/Collectible.tsx
