+++ apps/mobile/src/stacks/Explorer/index.tsx
@@ -3,8 +3,8 @@ import { withStackContainer } from '@walless/app';
 import { runtime } from '@walless/core';
 import type { DrawerNavigationOptions } from 'components/DrawerNavigation';
 import { createDrawerNavigator } from 'components/DrawerNavigation';
-import CollectibleScreen from 'features/Widget/Collectible';
-import CollectionScreen from 'features/Widget/Collection';
+import CollectibleScreen from 'screens/Dashboard/Collectible';
+import CollectionScreen from 'screens/Dashboard/Collection';
 import WidgetScreen from 'screens/Dashboard/Widget';
 import type { ExploreParamList } from 'utils/navigation';
